requirements.txt
paho-mqtt
requests
PyYAML
